+++ Editor/Builder/BuilderSelection.cs
@@ -193,16 +193,25 @@ namespace Unity.UI.Builder
                 var vta = m_Builder.document.visualTreeAsset;
                 var rule = vta.GetOrCreateInlineStyleRule(vea);
 
+#if UNITY_2020_1_OR_NEWER
+                element.SetInlineRule(vta.inlineSheet, rule);
+
+                // Need to enforce this specific style is updated.
+                element.IncrementVersion(VersionChangeType.Opacity);
+
+#elif UNITY_2019_3_OR_NEWER
                 var stylesData = new UnityEngine.UIElements.StyleSheets.VisualElementStylesData(false);
                 element.m_Style = stylesData;
-
-#if UNITY_2019_3_OR_NEWER
+                
                 s_StylePropertyReader.SetInlineContext(vta.inlineSheet, rule, vea.ruleIndex);
                 stylesData.ApplyProperties(s_StylePropertyReader, null);
 
                 // Need to enforce this specific style is updated.
                 element.IncrementVersion(VersionChangeType.Opacity);
 #else
+                var stylesData = new UnityEngine.UIElements.StyleSheets.VisualElementStylesData(false);
+                element.m_Style = stylesData;
+
                 var propIds = UnityEngine.UIElements.StyleSheets.StyleSheetCache.GetPropertyIDs(vta.inlineSheet, vea.ruleIndex);
                 element.specifiedStyle.ApplyRule(vta.inlineSheet, Int32.MaxValue, rule, propIds);
 #endif